int -> size_t

+++ lang/c++/api/Node.hh
@@ -122,8 +122,8 @@ class AVRO_DECL Node : private boost::noncopyable
         doAddLeaf(newLeaf);
     }
     virtual size_t leaves() const = 0;
-    virtual const NodePtr& leafAt(int index) const = 0;
-    virtual const GenericDatum& defaultValueAt(int index) {
+    virtual const NodePtr& leafAt(size_t index) const = 0;
+    virtual const GenericDatum& defaultValueAt(size_t index) {
         throw Exception(boost::format("No default value at: %1%") % index);
     }
 
@@ -133,14 +133,14 @@ class AVRO_DECL Node : private boost::noncopyable
         doAddName(name);
     }
     virtual size_t names() const = 0;
-    virtual const std::string &nameAt(int index) const = 0;
+    virtual const std::string &nameAt(size_t index) const = 0;
     virtual bool nameIndex(const std::string &name, size_t &index) const = 0;
 
     void setFixedSize(int size) {
         checkLock();
         doSetFixedSize(size);
     }
-    virtual int fixedSize() const = 0;
+    virtual size_t fixedSize() const = 0;
 
     virtual bool isValid() const = 0;
 
@@ -150,7 +150,7 @@ class AVRO_DECL Node : private boost::noncopyable
 
     virtual void printBasicInfo(std::ostream &os) const = 0;
 
-    virtual void setLeafToSymbolic(int index, const NodePtr &node) = 0;
+    virtual void setLeafToSymbolic(size_t index, const NodePtr &node) = 0;
 
   protected:
 
@@ -167,7 +167,7 @@ class AVRO_DECL Node : private boost::noncopyable
     virtual void doSetName(const Name &name) = 0;
     virtual void doAddLeaf(const NodePtr &newLeaf) = 0;
     virtual void doAddName(const std::string &name) = 0;
-    virtual void doSetFixedSize(int size) = 0;
+    virtual void doSetFixedSize(size_t size) = 0;
 
   private:
